# wso2/apextick-api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdmissionsResponse": {
        "properties": {
          "admitted": {
            "format": "int64",
            "type": "integer"
          },
          "eventId": {
            "format": "int64",
            "type": "integer"
          },
          "issued": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "CurrentUser": {
        "properties": {
          "admin": {
            "type": "boolean"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true
          },
          "sub": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "GateEventResponse": {
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "enum": [
              "onsale",
              "selling-fast",
              "final-release",
              "sold-out",
              "draft",
              "cancelled"
            ],
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "venue": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ScanRequest": {
        "properties": {
          "eventId": {
            "format": "int64",
            "type": "integer"
          },
          "gate": {
            "maxLength": 64,
            "minLength": 0,
            "type": "string"
          },
          "qrToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "qrToken"
        ],
        "type": "object"
      },
      "ScanResponse": {
        "properties": {
          "admissions": {
            "$ref": "#/components/schemas/AdmissionsResponse"
          },
          "event": {
            "$ref": "#/components/schemas/GateEventResponse"
          },
          "ticket": {
            "$ref": "#/components/schemas/TicketResponse"
          }
        },
        "type": "object"
      },
...
      "TicketResponse": {
        "properties": {
          "eventId": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "issuedAt": {
            "format": "date-time",
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "pdfUrl": {
            "type": "string"
          },
          "qrToken": {
            "type": "string"
          },
          "seatId": {
            "format": "int64",
            "type": "integer"
          },
          "seatLabel": {
            "type": "string"
          },
          "sectionName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tierName": {
            "type": "string"
          },
          "usedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UnadmitRequest": {
        "properties": {
          "reason": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
      "UnadmitResponse": {
        "properties": {
          "admissions": {
            "$ref": "#/components/schemas/AdmissionsResponse"
          },
          "ticket": {
            "$ref": "#/components/schemas/TicketResponse"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for browsing events and holding/booking seats.",
    "title": "ApexTick Booking & Inventory API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/gate/events/{eventId}/admissions": {
      "get": {
        "operationId": "admissions",
        "parameters": [
          {
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionsResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Tickets admitted to an event so far, across every gate",
        "tags": [
          "Gate"
        ]
      }
    },
    "/gate/scans": {
      "post": {
        "operationId": "scan",
        "parameters": [
          {
            "in": "query",
            "name": "steward",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CurrentUser"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Admit the ticket behind a scanned QR code to the event this gate is admitting",
        "tags": [
          "Gate"
        ]
      }
    },
    "/gate/tickets/{ticketId}/unadmit": {
      "post": {
        "operationId": "unadmit",
        "parameters": [
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "admin",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CurrentUser"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnadmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnadmitResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Undo a mistaken admission, with a reason for the record",
        "tags": [
          "Gate"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}